+++ src/Commands/Concerns/InteractsWithServers.php
@@ -107,6 +107,19 @@ trait InteractsWithServers
     }
 
     /**
+     * Retrieve the given server output and flush it.
+     *
+     * @return array
+     */
+    protected function getServerOutput($server)
+    {
+        return tap([
+            $server->getIncrementalOutput(),
+            $server->getIncrementalErrorOutput(),
+        ], fn () => $server->clearOutput()->clearErrorOutput());
+    }
+
+    /**
      * Returns the list of signals to subscribe.
      *
      * @return array